# Slim down the benchmark suite