--- a/clients/CMakeLists.txt
+++ b/clients/CMakeLists.txt
@@ -107,6 +107,7 @@ if(BUILD_CLIENTS_BENCHMARKS OR BUILD_CLIENTS_TESTS)
   set(common_source_files
     common/lapack_host_reference.cpp
     rocblascommon/clients_utility.cpp
+    rocblascommon/program_options.cpp
     ${explicit_inst_files}
   )