These changes are designed to reduce the runtime of our benchmark suite:

# Trim benchmark suite problems

- Reduce maximum problem sizes
  This will reduce overall runtime as well as the memory required to to run the suite. The largest getri_strided_batched benchmarks used >16 GiB of memory, which was too much for some GPUs.
- Reduce number of problems
  The getri benchmarks took significantly longer than the others, so the increment between matrix sizes was increased to reduce the  overall runtime of the benchmark suite. The geqrf tests and various strided_batched tests had smaller contributions to the overall runtime than getri, but were still significant and were reduced too.
- Simplify the use of chained ranges
  Python's range includes values in [min, max). This commit changes the usage from

      range(a, b+1, inc1), range(b+inc1, c, inc2)

  to  

      range(a, b, inc1), range(b, c, inc2)

  The only difference is that `b` has been moved from the first range to the second range. This is important for the strided_batched variants, as we use the same batch count for all elements in a range. For example, the first range may have a batch count of 5000 and the second may have a range of 2500. Since `b` has moved from the first range to the second range, it will have a lower batch count.

# Move regex initialization into an implementation file

Since the regex  object was marked static, each translation unit that included it got  its own copy. This resulted in dozens of copies of the regex being initialized (and optimized) during program initialization before calling into main.

For example, the first few lines of: 

    ltrace -cC rocsolver-bench -f geqrf_strided_batched -r s -m 48 --iters 10 --batch_count 5000

Before
------
```
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 30.98    3.405386      283782        12 rocsolver_sgeqrf_strided_batched
 18.07    1.985818          48     40685 operator new(unsigned long)
 14.89    1.636944          61     26520 memcpy
 11.74    1.290850      107570        12 hipMemcpy
  8.32    0.914766          31     29441 std::ctype<char> const& std::use_facet<std::ctype<char> >(std::locale const&)
  5.48    0.602188          34     17665 operator delete(void*)
  4.15    0.455793          31     14691 std::__cxx11::collate<char> const& std::use_facet<std::__cxx11::collate<char> >(std::locale const&)
  1.66    0.182649          57      3201 memmove
```
After
-----
```
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 39.55    3.257511      271459        12 rocsolver_sgeqrf_strided_batched
 19.06    1.570270          59     26522 memcpy
 17.56    1.446569          57     25258 operator new(unsigned long)
 14.34    1.181109       98425        12 hipMemcpy
  2.03    0.167594          62      2679 memmove
  1.67    0.137848        6892        20 hipStreamSynchronize
  1.62    0.133081      133081         1 hipGetDeviceCount
  1.46    0.119986          53      2238 operator delete(void*)
```

I kind of want to further trim that. While I managed to eliminate 38% of the allocations with that change, I suspect the complexity of the option parsing still hides a few (smaller) needless inefficiencies. I'm tempted to refactor it towards something simpler.