@cgmb 2022-11-06
As it stands, my changes reduce the overall test suite time to ~16 hours on my 3950X workstation with a Radeon VII.

## Time in Minutes Spent on Benchmarks

```
time | function
-----|-----------------------
  18 | sgeqrf
  12 | sgeqrf_strided_batched
  19 | sgetrf
  13 | sgetrf_strided_batched
  49 | sgetri
  32 | sgetri_strided_batched
  22 | dgeqrf
  13 | dgeqrf_strided_batched
  20 | dgetrf
  15 | dgetrf_strided_batched
  57 | dgetri
  34 | dgetri_strided_batched
  31 | cgeqrf
  21 | cgeqrf_strided_batched
  49 | cgetrf
  27 | cgetrf_strided_batched
 108 | cgetri
  66 | cgetri_strided_batched
  39 | zgeqrf
  24 | zgeqrf_strided_batched
  53 | zgetrf
  32 | zgetrf_strided_batched
 119 | zgetri
 dnf | zgetri_strided_batched
```

That's an estimate, though, since there was a crash in the final getri_strided_batched.

```
PATH="$HOME/rocSOLVER/build/release/clients/staging:$PATH" rocsolver-bench -f getri_strided_batched -r z -n 232 --iters 10 --batch_count 2500
rocSOLVER version 3.21.0.030d09af (with rocBLAS 2.47.0.dba9c1ba)
Query device success: there are 1 devices
-------------------------------------------------------------------------------
Device ID 0 : AMD Radeon VII
with 17.2 GB memory, max. SCLK 1801 MHz, max. MCLK 1000 MHz, compute capability 9.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 64
-------------------------------------------------------------------------------
Memory access fault by GPU node-1 (Agent handle: 0x198c320) on address 0x7fd16fad1000. Reason: Page not present or supervisor privilege.
Aborted (core dumped)
```

Those rocSOLVER / rocBLAS versions are pretty much just the current develop branch versions, so I think this indicates there's a bug in GETRI or its test setup.