@TorreZuk 2022-11-10
@RobsonRLemos based on these huge run times I think you need to up the priority for moving to an opt-in for PRs for perf runs to avoid huge energy waste.   @cgmb did you make a ticket for what looks like the int_32 overflow bug you found... could it be in rocBLAS or can say it is in rocSOLVER?