@cgmb 2022-11-10
> @cgmb did you make a ticket for what looks like the int_32 overflow bug you found... could it be in rocBLAS or can say it is in rocSOLVER?

It's recorded as SWDEV-367057. I suspect the problem is in rocsolver-bench, not in the rocblas or rocsolver libraries.