@RobsonRLemos 2022-11-10
@TorreZuk that is a good point. I will further discuss this with the CI system stakeholders how to set up the priorities for the perf runs.